+++ blunder/main.go
@@ -6,7 +6,7 @@ import (
 
 func main() {
 	/*var s engine.Search
-	s.Pos.LoadFEN("2r2rk1/1bqnbpp1/1p1ppn1p/pP6/N1P1P3/P2B1N1P/1B2QPP1/R2R2K1 b - - 0 1")
+	s.Pos.LoadFEN("8/6k1/8/8/4K3/8/8/8 w - - 2 190")
 	s.Timer.TimeLeft = 300000
 	s.TT.Resize(engine.DefaultTTSize)
 	start := time.Now()